update to jrtp 3.5.2 and jthread 1.2.0
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@954 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/jrtplib/src/rtpsessionparams.h
@@ -90,7 +90,7 @@ public:
 private:
 	bool acceptown;
 	bool usepollthread;
-	int maxpacksize;
+	size_t maxpacksize;
 	double owntsunit;
 	RTPTransmitter::ReceiveMode receivemode;
 	bool resolvehostname;
